Update file organization
